Refactor time commands to their own class
