+++ HandyHansel/Commands/GeneralCommands.cs
@@ -32,122 +32,5 @@ namespace HandyHansel.Commands
             if (!result.TimedOut)
                 await context.RespondAsync($"I'm fine, thank you!");
         }
-
-        [Command("time")]
-        public async Task CurrentTime(CommandContext context)
-        {
-            await context.RespondAsync($":wave: Hi, {context.User.Mention}! What timezone do you want the time for?");
-            InteractivityExtension interactivity = context.Client.GetInteractivity();
-            List<TimeZoneInfo> correctResponses = getListOfTimeZones();
-
-            IDictionary<int, DiscordEmoji> discordEmojiNumbers = new Dictionary<int, DiscordEmoji>
-            {
-                { 0, DiscordEmoji.FromName(context.Client, ":zero:") },
-                { 1, DiscordEmoji.FromName(context.Client, ":one:") },
-                { 2, DiscordEmoji.FromName(context.Client, ":two:") },
-                { 3, DiscordEmoji.FromName(context.Client, ":three:") },
-                { 4, DiscordEmoji.FromName(context.Client, ":four:") },
-                { 5, DiscordEmoji.FromName(context.Client, ":five:") },
-                { 6, DiscordEmoji.FromName(context.Client, ":six:") },
-                { 7, DiscordEmoji.FromName(context.Client, ":seven:") },
-                { 8, DiscordEmoji.FromName(context.Client, ":eight:") },
-                { 9, DiscordEmoji.FromName(context.Client, ":nine:") },
-                { 10, DiscordEmoji.FromName(context.Client, ":keycap_ten:") },
-            };
-
-            discordEmojiNumbers = discordEmojiNumbers.Where(keyValuePair => keyValuePair.Key < correctResponses.Count).ToDictionary(keyValuePair => keyValuePair.Key, keyValuePair => keyValuePair.Value);
-
-            string description = "";
-            for (int i = 0; i < correctResponses.Count; i++)
-            {
-                description += discordEmojiNumbers[i].ToString() + " " + correctResponses[i].Id + (i == correctResponses.Count - 1 ? " " : "\n\n");
-            }
-
-            DiscordEmbed embed = new DiscordEmbedBuilder
-            {
-                Title = "Choose a time!",
-                Description = description,
-            };
-
-            DiscordMessage msg = await context.RespondAsync(embed: embed);
-
-            for (int i = 0; i < correctResponses.Count && discordEmojiNumbers.ContainsKey(i); i++)
-            {
-                await msg.CreateReactionAsync(discordEmojiNumbers[i]);
-            }
-
-            InteractivityResult<DSharpPlus.EventArgs.MessageReactionAddEventArgs> result =
-                await interactivity.WaitForReactionAsync(reaction =>
-                    discordEmojiNumbers.Any(keyValuePair =>
-                    {
-                        return keyValuePair.Value.Equals(reaction.Emoji) && !reaction.User.IsBot && reaction.Message.Id == msg.Id;
-                    }
-                    )
-                );
-
-            if (!result.TimedOut)
-            {
-                int discordEmojiValue = discordEmojiNumbers.First(keyValuePair => keyValuePair.Value.Equals(result.Result.Emoji)).Key;
-                DateTime currentTime = DateTime.Now;
-                TimeZoneInfo requestedTimeZone = correctResponses[discordEmojiValue];
-                DateTime requestedTime = TimeZoneInfo.ConvertTime(currentTime, TimeZoneInfo.Local, requestedTimeZone);
-                DiscordEmbed timeEmbed = new DiscordEmbedBuilder
-                {
-                    Title = "Current Time: " + requestedTimeZone.Id,
-                    Description = $"{requestedTime}",
-                };
-                await context.RespondAsync(embed: timeEmbed);
-            }
-        }
-
-        [Command("allTimes")]
-        public async Task AllTimesOnSystem(CommandContext context)
-        {
-            if (context.User.Username != "Prof Doofenshmirtz") return;
-            List<TimeZoneInfo> correctResponses = TimeZoneInfo.GetSystemTimeZones().ToList();
-            string description = "";
-            for (int i = 0; i < correctResponses.Count; i++)
-            {
-                description += correctResponses[i].Id + (i == correctResponses.Count - 1 ? " " : "\n\n");
-            }
-
-            System.IO.File.WriteAllText(@"AllTimes.txt", description);
-
-            await context.RespondWithFileAsync(@"AllTimes.txt");
-        }
-
-        [Command("currentDbTimes")]
-        public async Task CurrentTimeZonesOnDb(CommandContext context)
-        {
-            if (context.User.Username != "Prof Doofenshmirtz") return;
-            List<GuildTimeZone> allGuildTimeZones = DataAccessProvider.GetAllGuildsTimeZones();
-            string description = "Guild,TimeZoneId,OperatingSystem\n";
-            for (int i = 0; i < allGuildTimeZones.Count; i++)
-            {
-                description += allGuildTimeZones[i].Guild + "," + allGuildTimeZones[i].TimeZoneId + "," + allGuildTimeZones[i].OperatingSystem +  (i == allGuildTimeZones.Count - 1 ? " " : "\n");
-            }
-
-            System.IO.File.WriteAllText(@"CurrentDbTimes.csv", description);
-
-            await context.RespondWithFileAsync(@"CurrentDbTimes.csv");
-        }
-
-        List<TimeZoneInfo> getListOfTimeZones()
-        {
-            List<TimeZoneInfo> timeZoneInfos = new List<TimeZoneInfo>();
-            System.Collections.ObjectModel.ReadOnlyCollection<TimeZoneInfo> allTimeZones = TimeZoneInfo.GetSystemTimeZones();
-            foreach (TimeZoneInfo timezone in TimeZoneInfo.GetSystemTimeZones())
-            {
-                if (timezone.Id.Equals("Central Standard Time") // Windows CST
-                 || timezone.Id.Equals("AUS Eastern Standard Time") // Windows AEST
-                 || timezone.Id.Equals("America/Chicago") // Ubuntu America/Chicago
-                 || timezone.Id.Equals("America/New York") // Ubuntu America/New York
-                 || timezone.Id.Equals("Australia/Sydney") // Ubuntu Australia/Sydney
-                 || timezone.Id.Equals("Australia/Melbourne") // Ubuntu Australia/Melbourne
-                )
-                    timeZoneInfos.Add(timezone);
-            }
-            return timeZoneInfos;
-        }
     }
 }
