+++ HandyHansel/HandyHansel.csproj
@@ -6,14 +6,14 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="DSharpPlus" Version="4.0.0-nightly-00694" />
-    <PackageReference Include="DSharpPlus.CommandsNext" Version="4.0.0-nightly-00694" />
-    <PackageReference Include="DSharpPlus.Interactivity" Version="4.0.0-nightly-00694" />
-    <PackageReference Include="DSharpPlus.Lavalink" Version="4.0.0-nightly-00694" />
-    <PackageReference Include="DSharpPlus.Rest" Version="4.0.0-nightly-00694" />
-    <PackageReference Include="DSharpPlus.VoiceNext" Version="4.0.0-nightly-00694" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.1.4" />
-    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="3.1.3" />
+    <PackageReference Include="DSharpPlus" Version="4.0.0-nightly-00697" />
+    <PackageReference Include="DSharpPlus.CommandsNext" Version="4.0.0-nightly-00697" />
+    <PackageReference Include="DSharpPlus.Interactivity" Version="4.0.0-nightly-00697" />
+    <PackageReference Include="DSharpPlus.Lavalink" Version="4.0.0-nightly-00697" />
+    <PackageReference Include="DSharpPlus.Rest" Version="4.0.0-nightly-00697" />
+    <PackageReference Include="DSharpPlus.VoiceNext" Version="4.0.0-nightly-00697" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="5.0.0-preview.3.20181.2" />
+    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="5.0.0-preview3" />
   </ItemGroup>
 
 </Project>
